requirements.txt
# Core dependencies
pydantic>=2.0.0

# Graph database
neo4j>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Vision model
transformers>=4.37.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0

# Hugging Face Inference API
huggingface_hub>=0.20.0

